refactor(doughnutChart): seperate child component, Sector.
