+++ src/features/Statistics/DoughnutChart.tsx
@@ -4,7 +4,7 @@ import {
   faCaretRight,
 } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import React, { ReactNode } from 'react';
+import React from 'react';
 import styled from 'styled-components/macro';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
 import {
@@ -12,13 +12,12 @@ import {
   CHOOSE_MONTH,
 } from '../../redux/reducers/ledgerListSlice';
 import { SWITCH_STATISTICS_MODE } from '../../redux/reducers/pageControlSlice';
+import { DoughnutSector } from './DoughnutSector';
 
 interface Props {
-  props: {
-    setting: ChartSetting;
-    data: { label: string; value: number; colorCode: string }[];
-    title: string[];
-  };
+  setting: ChartSetting;
+  data: { label: string; value: number; colorCode: string }[];
+  title: string[];
 }
 
 interface ChartSetting {
@@ -28,7 +27,7 @@ interface ChartSetting {
   labelDelta: number;
 }
 
-export const DoughnutChart: React.FC<Props> = ({ props }) => {
+export const DoughnutChart: React.FC<Props> = ({ setting, data, title }) => {
   const { chosenLabel, chosenMonth } = useAppSelector(
     (state) => state.ledgerList.choices
   );
@@ -36,8 +35,7 @@ export const DoughnutChart: React.FC<Props> = ({ props }) => {
     (state) => state.pageControl
   );
   const dispatch = useAppDispatch();
-  const { radius, holeRatio, hoverDelta, labelDelta } = props.setting;
-  const { data, title } = props;
+  const { radius, hoverDelta, labelDelta } = setting;
 
   const bufferSpace = 20;
   const abstractSpace = 50;
@@ -50,7 +48,7 @@ export const DoughnutChart: React.FC<Props> = ({ props }) => {
     totalValue === 0 ? 0 : data.value / totalValue
   );
 
-  const startRatios = data.map((data, dataIndex) =>
+  const startRatios = data.map((_, dataIndex) =>
     dataIndex === 0
       ? 0
       : sectorRatios.reduce(
@@ -72,128 +70,19 @@ export const DoughnutChart: React.FC<Props> = ({ props }) => {
     };
   });
 
-  const drawSector = (
-    data: {
-      sectorRatio: number;
-      startAngle: number;
-      label: string;
-      value: number;
-      colorCode: string;
-    },
-    index: number
-  ): ReactNode => {
-    const { sectorRatio, startAngle, label, colorCode } = data;
-
-    const origin = {
-      x: svgWidth / 2,
-      y: svgHeight / 2,
-    };
-    const midAngle = startAngle + sectorRatio * Math.PI;
-    const endAngle = startAngle + sectorRatio * 2 * Math.PI;
-    const startPoint = {
-      x: radius * Math.cos(startAngle) + origin.x,
-      y: radius * Math.sin(startAngle) + origin.y,
-    };
-    const endPoint = {
-      x: radius * Math.cos(endAngle) + origin.x,
-      y: radius * Math.sin(endAngle) + origin.y,
-    };
-    const holeRadius = radius * holeRatio;
-    const angleAdjust = 0.01;
-    const holeStartPoint = {
-      x: holeRadius * Math.cos(startAngle - angleAdjust) + origin.x,
-      y: holeRadius * Math.sin(startAngle - angleAdjust) + origin.y,
-    };
-    const holeEndPoint = {
-      x: holeRadius * Math.cos(endAngle + angleAdjust) + origin.x,
-      y: holeRadius * Math.sin(endAngle + angleAdjust) + origin.y,
-    };
-    const largeArcFlag = sectorRatio > 0.5 ? 1 : 0;
-
-    const colorDScript = `M ${startPoint.x} ${startPoint.y} A ${radius} ${radius} 0 ${largeArcFlag} 1 ${endPoint.x} ${endPoint.y} L ${origin.x} ${origin.y} Z`;
-    const holeDScript = `M ${holeStartPoint.x} ${holeStartPoint.y} A ${holeRadius} ${holeRadius} 0 ${largeArcFlag} 1 ${holeEndPoint.x} ${holeEndPoint.y} L ${origin.x} ${origin.y} Z`;
-
-    const labelNameAnchor =
-      sectorRatio > 0.05
-        ? {
-            x: (radius - labelDelta) * Math.cos(midAngle) + origin.x,
-            y: (radius - labelDelta) * Math.sin(midAngle) + origin.y,
-          }
-        : {
-            x: (radius + 0.15 * labelDelta) * Math.cos(midAngle) + origin.x,
-            y: (radius + 0.15 * labelDelta) * Math.sin(midAngle) + origin.y,
-          };
-    const labelRatioAnchor = {
-      x: (radius + labelDelta) * Math.cos(midAngle) + origin.x,
-      y: (radius + labelDelta) * Math.sin(midAngle) + origin.y,
-    };
-
-    const hoverShift = {
-      x: hoverDelta * Math.cos(midAngle),
-      y: hoverDelta * Math.sin(midAngle),
-    };
-
-    const handleClickingLabel = () => {
-      label === chosenLabel
-        ? dispatch(CHOOSE_LABEL(''))
-        : dispatch(CHOOSE_LABEL(label));
-      if (chartShown !== 'monthAndDetail')
-        dispatch(SWITCH_STATISTICS_MODE('monthAndDetail'));
-    };
-
-    return (
-      <SectorGroup
-        key={index}
-        $xShift={hoverShift.x}
-        $yShift={hoverShift.y}
-        $isChosen={label === chosenLabel}
-      >
-        {sectorRatio === 1 ? (
-          <>
-            <CirclePath
-              onClick={handleClickingLabel}
-              cx={origin.x}
-              cy={origin.y}
-              r={radius}
-              fill={colorCode}
-            />
-            <CircleHolePath
-              cx={origin.x}
-              cy={origin.y}
-              r={holeRadius}
-              fill={'#f2f2f2'}
-            />
-          </>
-        ) : (
-          <>
-            <SectorPath
-              onClick={handleClickingLabel}
-              d={colorDScript}
-              fill={colorCode}
-            />
-            <HolePath d={holeDScript} fill={'#f2f2f2'} />
-          </>
-        )}
-
-        <LabelName
-          x={labelNameAnchor.x}
-          y={labelNameAnchor.y}
-          onClick={handleClickingLabel}
-        >
-          {label}
-        </LabelName>
-        <LabelRatio x={labelRatioAnchor.x} y={labelRatioAnchor.y}>
-          {`${Math.round(sectorRatio * 100)}%`}
-        </LabelRatio>
-      </SectorGroup>
-    );
-  };
-
   return (
     <ChartWrap>
       <PieSvg $svgHeight={svgHeight} $svgWidth={svgWidth}>
         {conbinedData.map((data, index) =>
-          data.sectorRatio === 0 ? '' : drawSector(data, index)
+          data.sectorRatio === 0 ? (
+            ''
+          ) : (
+            <DoughnutSector
+              data={data}
+              index={index}
+              setting={{ ...setting, svgWidth, svgHeight }}
+            />
+          )
         )}
         <ChartSubtitle x={svgWidth / 2 - 70} y={(abstractSpace * 2) / 3}>
           {title[0]}
@@ -263,12 +152,6 @@ type PieSvgProps = {
   $svgWidth: number;
 };
 
-type SectorGroupProps = {
-  $xShift: number;
-  $yShift: number;
-  $isChosen: boolean;
-};
-
 type ChartAbstractProps = {
   $color: string;
 };
@@ -287,49 +170,6 @@ const PieSvg = styled.svg<PieSvgProps>`
   padding-top: 0px;
   height: ${({ $svgHeight }) => `${$svgHeight}px`};
   width: ${({ $svgWidth }) => `${$svgWidth}px`};
-`;
-
-const SectorGroup = styled.g<SectorGroupProps>`
-  transition: transform 0.6s ease;
-  opacity: 0.7;
-  transform: ${({ $xShift, $yShift, $isChosen }) =>
-    $isChosen ? `translate(${$xShift}px, ${$yShift}px)` : ''};
-  &:hover {
-    transform: ${({ $xShift, $yShift }) =>
-      `translate(${$xShift}px, ${$yShift}px)`};
-  }
-`;
-
-const SectorPath = styled.path`
-  cursor: pointer;
-`;
-const HolePath = styled(SectorPath)`
-  cursor: default;
-`;
-
-const CirclePath = styled.circle`
-  cursor: pointer;
-`;
-const CircleHolePath = styled(CirclePath)`
-  cursor: default;
-`;
-
-const LabelName = styled.text`
-  font-size: 16px;
-  font-weight: bold;
-  text-anchor: middle;
-
-  fill: #5b4105;
-
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
-const LabelRatio = styled.text`
-  font-size: 20px;
-  text-anchor: middle;
-  fill: #5b4105;
 `;
 
 const ChartTitle = styled.text`
